Container image caching

+++ BrainPortal/app/models/cluster_task.rb
@@ -2345,6 +2345,8 @@ chmod 755 #{singularity_wrapper_basename.bash_escape}
   private
 
   # Give a unique name for container image
+  # This is used as an entry in the task's work directory, usually
+  # as a symbolic link to the real content of the image.
   def container_image_name #:nodoc:
     ".container-#{self.id}.img"
   end
@@ -2365,32 +2367,53 @@ chmod 755 #{singularity_wrapper_basename.bash_escape}
     singularity_image = self.tool_config.container_image
     self.addlog("Syncing the singularity image '#{singularity_image.name}'")
 
-    image_name = container_image_name
+    # Sync the userfile content
     singularity_image.sync_to_cache
+
+    # Create the symlink to the cached image.
     cachename = singularity_image.cache_full_path
+    image_name = container_image_name
     safe_symlink(cachename,image_name)
   end
 
-  # Perform the singularity pull
+  # Perform the singularity pull; the image will be cached
+  # as a special, hidden userfile on the ScratchDataProvider.
   def load_singularity_image_from_repo #:nodoc:
-    singularity_image_name = self.tool_config.containerhub_image_name
+    singularity_image_name     = self.tool_config.containerhub_image_name
     singularity_index_location = self.tool_config.container_index_location.presence || "shub://"
 
     self.addlog("Pulling singularity image '#{singularity_image_name}'")
 
+    # Find or create the userfile holding the image content.
+    scratch_name     = "Singularity-pull-" + singularity_image_name.gsub(/[^a-z0-9_\.\-]+/i,"_") # must respect userfile convention.
+    scratch_name.sub!(/(\.img)?$/i, ".img") # singularity pull is peculiar about extensions... :-(
+    scratch_userfile = SingularityImage.find_or_create_as_scratch(:name => scratch_name) do |cache_path|
+      # Optimization: if another find_or_create_as_scratch has already beaten us to the punch
+      # and dowloaded the file, just skip this block altogether. The way SyncStatus works, several
+      # of these blocks can be scheduled to run, but only one will execute at at any given time.
+      next if File.exists?(cache_path.to_s) && File.size(cache_path.to_s) > 0
+      # Run singularity pull command
+      errfile = "/tmp/.container_load_cmd.#{self.run_id}.err"
+      success = Dir.chdir(cache_path.parent.to_s) do # singularity pull won't work with full pathnames for --name :-(
+        system("#{singularity_executable_name} pull --name #{scratch_name.bash_escape} #{singularity_index_location.bash_escape}#{singularity_image_name.bash_escape} </dev/null >/dev/null 2>#{errfile.bash_escape}")
+      end
+      err     = File.read(errfile) rescue "No Error File?"
+      File.unlink(errfile) rescue true
+      # Singularity command can generate 'implausibly old time stamp' when pulling a docker image (due to tar), we ignore it.
+      # Remove all lines (use ^ and $) that contains this message.
+      err.gsub!(/^.*implausibly old time stamp.*$\n?/,"")
+      cb_error "Cannot pull singularity image" if
+        err.present? ||
+        ! success    ||
+        ! File.exists?(cache_path.to_s)
+    end
+
+    self.addlog("Image stored as scratch userfile '#{scratch_userfile.name}' (ID #{scratch_userfile.id})")
+
+    # Create the symlink to the cached image.
+    cachename  = scratch_userfile.cache_full_path
     image_name = container_image_name
-    errfile = "/tmp/.container_load_cmd.#{self.run_id}.err"
-    success = system("#{singularity_executable_name} pull --name #{image_name.bash_escape} #{singularity_index_location.bash_escape}#{singularity_image_name.bash_escape} </dev/null >/dev/null 2>#{errfile.bash_escape}")
-    err     = File.read(errfile) rescue "No Error File?"
-    # Singularity command can generate 'implausibly old time stamp' when pulling a docker image (due to tar), we ignore it.
-    # Remove all lines (use ^ and $) that contains this message. 
-    err.gsub!(/^.*implausibly old time stamp.*$\n?/,"")
-    cb_error "Cannot pull singularity image" if
-      err.present? ||
-      ! success    ||
-      ! File.exists?(image_name)
-  ensure
-    File.unlink(errfile) rescue true
+    safe_symlink(cachename,image_name)
   end
 
   ##################################################################
